% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_CIS.R
\name{calculate_CIS}
\alias{calculate_CIS}
\title{Calculate Composite Interaction Score (CIS)}
\usage{
calculate_CIS(df, p = 0.9)
}
\arguments{
\item{df}{Data frame with common interactions}

\item{p}{Patience parameter (default = 0.9)}
}
\value{
Data frame with CIS scores and rankings
}
\description{
Calculate Composite Interaction Score (CIS)
}
