% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_results.R
\name{combine_results}
\alias{combine_results}
\title{Combine results from different LIANA methods}
\usage{
combine_results(cellchat, cellphonedb, connectome, logfc, natmi, sca)
}
\arguments{
\item{...}{CSV file paths of each method's output}
}
\value{
A data frame of interactions common to all methods
}
\description{
Combine results from different LIANA methods
}
