% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cellchat.R
\name{run_cellchat}
\alias{run_cellchat}
\title{Run LIANA with CellChat method}
\usage{
run_cellchat(seurat_obj, idents_col = "celltype.combo")
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{idents_col}{Column for cell types}
}
\value{
Data frame of inferred CCIs
}
\description{
Run LIANA with CellChat method
}
