% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cellphonedb.R
\name{run_cellphonedb}
\alias{run_cellphonedb}
\title{Run LIANA with CellPhoneDB-style permutation}
\usage{
run_cellphonedb(seurat_obj, idents_col = "celltype.combo")
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{idents_col}{Column for cell types}
}
\value{
Data frame of inferred CCIs
}
\description{
Run LIANA with CellPhoneDB-style permutation
}
