% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_liana.R
\name{run_liana}
\alias{run_liana}
\title{Run LIANA with multiple methods}
\usage{
run_liana(seurat_obj, idents_col = "celltype.combo")
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{idents_col}{Column name in metadata for identities}
}
\value{
A named list of data frames from different methods
}
\description{
Run LIANA with multiple methods
}
